Updated content for SEIR-129
Reworked the examples in codepen and replaced the older ones to
simplify the concepts and use GA branding and color palette.
Reworked the content to have a less negative introduction to CSS.
CSS is wonderful and telling people that it isn't or that it's hard
and devs find it frustrating will naturally color their experience
with it.  It's a bad message.  Also changed up the content so that
it follows an I Do, We Do, You Do model.